requirements.txt
streamlit
pandas
google-generativeai
Pillow
openpyxl
xlsxwriter